refactor(lvgl): split scene bindings into focused modules
